Fixed return type in Audio.getRating().

+++ src/com/github/af0liveira/aluraplay/modelos/Audio.java
@@ -48,7 +48,7 @@ public class Audio {
         return this.timesDisliked;
     }
 
-    public float getRating() {
+    public int getRating() {
         float scalingFactor 
             = (float) this.timesLiked / (this.timesLiked + this.timesDisliked);
         int rating = (int) (scalingFactor * this.timesPlayed);
